html file
This HTML file structures the content of my resume website, featuring sections for skills, experience, education, and projects. Built with a clean and semantic approach, it ensures easy readability and maintainability. The file is linked to an external CSS file for styling, creating a modern and responsive design. This project highlights my ability to create well-organized and visually appealing web pages using HTML.